% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbsDB_parse_and_plot.R
\name{merge_same_tf}
\alias{merge_same_tf}
\title{Merge overlapping TFBS}
\usage{
merge_same_tf(gr, minfrac = 0.05, fxn = min)
}
\arguments{
\item{gr}{GRange object of TFBS}

\item{minfrac}{minimum overlap between to windows to merge TFBS's}

\item{fxn}{score of merged interval is the assigned as fxn(score(.)) of overlapping intervals}
}
\description{
Merge adjacent TFBS that overlap by at least minfrac
}
