% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbsDB_parse_and_plot.R
\name{mergeOverlapping}
\alias{mergeOverlapping}
\title{Merge overlapping GRanges entries}
\usage{
mergeOverlapping(x, y, minfrac = 0.05)
}
\arguments{
\item{x}{GRange object}

\item{y}{GRange object}

\item{minfrac}{minimum overlap between to windows to merge entries}
}
\description{
Merge adjacent GRanges entries that overlap by at least minfrac
}
