% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbsDB_parse_and_plot.R
\name{readJaspar}
\alias{readJaspar}
\title{Read BED file of TFBS scores from JASPAR into GRange object}
\usage{
readJaspar(grQuery, bedFile, maxP = 1e-04)
}
\arguments{
\item{grQuery}{GRange object of query interval}

\item{bedFile}{BED file that is bgzipped and tabix'd.}

\item{maxP}{maximum p-value of TFBS returned. score corresponds to p-value: 200 -> 10e-2; 300 -> 10e-3}
}
\value{
GRange of TFBS locations where scores are transformed p-values
}
\description{
Read BED file of TFBS scores from JASPAR into GRange object
}
\examples{
library(GenomicRanges)
library(tfbsDB)

# range of TFBS to be loaded from file
grQuery = GRanges("chr1", IRanges(10280, 10410))

# example data from package
bedFile = file.path(system.file("inst/extdata/", package = "tfbsDB"), "JASPAR2018_hg19_small.bed.gz")

# read from file
gr = readJaspar( grQuery, bedFile )

# show TFBS locations
plotTFBSdb( gr )

# color locations by p-value of motif match
plotTFBSdb( gr, colorByP=TRUE)

}
