% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbsDB_parse_and_plot.R
\name{plotTFBSdb}
\alias{plotTFBSdb}
\title{Plot TFBS in window}
\usage{
plotTFBSdb(gr, xlim = c(start(gr), end(gr)), aspect.ratio = 0.1,
  tf_text_size = 6, merge_tfbs = TRUE, merge_min_frac = 0.05,
  segmentColor = "lightblue", textColor = "black", colorByP = FALSE,
  gradientRange = c(4, 10))
}
\arguments{
\item{gr}{GRange object of TFBS}

\item{xlim}{range of window in genome coordinates}

\item{aspect.ratio}{aspect ratio of plot}

\item{tf_text_size}{size of TF text label}

\item{merge_tfbs}{should overlapping TFBS from same TF be combined?}

\item{merge_min_frac}{minimum overlap between to windows to merge TFBS's}

\item{segmentColor}{color of TFBS segments}

\item{textColor}{color of TFBS text labels}

\item{colorByP}{color segments by -log10(p) of motif match}

\item{gradientRange}{ranges of values for colors when colorByP is TRUE}
}
\description{
Produce ggbio plot that ca be combined with others
}
