% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbsDB_parse_and_plot.R
\name{plotTFBSdb}
\alias{plotTFBSdb}
\title{Plot TFBS in window}
\usage{
plotTFBSdb(gr, xlim = c(min(start(gr)), max(end(gr))),
  tf_text_size = 6, merge_tfbs = TRUE, merge_min_frac = 0.05,
  segmentColor = "lightblue", textColor = "black", colorByP = FALSE,
  gradientRange = c(4, 10))
}
\arguments{
\item{gr}{GRange object of TFBS}

\item{xlim}{range of window in genome coordinates}

\item{tf_text_size}{size of TF text label}

\item{merge_tfbs}{should overlapping TFBS from same TF be combined?}

\item{merge_min_frac}{minimum overlap between to windows to merge TFBS's}

\item{segmentColor}{color of TFBS segments}

\item{textColor}{color of TFBS text labels}

\item{colorByP}{color segments by -log10(p) of motif match}

\item{gradientRange}{ranges of values for colors when colorByP is TRUE}
}
\value{
plot as ggbio object
}
\description{
Produce ggbio plot that ca be combined with others
}
\examples{
library(GenomicRanges)
library(tfbsDB)

# range of TFBS to be loaded from file
grQuery = GRanges("chr1", IRanges(10280, 10410))

# example data from package
bedFile = file.path(system.file("extdata/", package = "tfbsDB"), "hocomocov11_hg19_example.bed.gz")

# read from file
gr = readTFBSdb( grQuery, bedFile )

# show TFBS locations
plotTFBSdb( gr )

# color locations by p-value of motif match
plotTFBSdb( gr, colorByP=TRUE)

}
