% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbsDB_parse_and_plot.R
\name{readTFBSdb}
\alias{readTFBSdb}
\title{Read BED file of TFBS scores into GRange object}
\usage{
readTFBSdb(grQuery, bedFile, maxP = 1e-04, quality = c("A", "B", "C"))
}
\arguments{
\item{grQuery}{GRange object of query interval}

\item{bedFile}{BED file that is bgzipped and tabix'd.}

\item{maxP}{maximum p-value of TFBS returned}

\item{quality}{each motif has a quality value in c('A', 'B', 'C', 'D').  But default exclude quality D, the lowest quality motifs}
}
\value{
GRange of TFBS locations where p-values are scores
}
\description{
Read BED file of TFBS scores into GRange object
}
\examples{
library(GenomicRanges)
library(tfbsDB)

# range of TFBS to be loaded from file
grQuery = GRanges("chr1", IRanges(10280, 10410))

# example data from package
bedFile = file.path(system.file("data", package = "tfbsDB"), "motifs_fimo_exmaple.bed.gz")

# read from file
gr = readTFBSdb( grQuery, bedFile )

# show TFBS locations
plotTFBSdb( gr )

# color locations by p-value of motif match
plotTFBSdb( gr, colorByP=TRUE)

}
