% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbsDB_parse_and_plot.R
\name{readTFBSdb}
\alias{readTFBSdb}
\title{Read BED file of TFBS scores into GRange object}
\usage{
readTFBSdb(grQuery, bedFile = file.path(system.file("data", package =
  "tfbsDB"), "motifs_fimo_search_small.bed.gz"), maxP = 1e-04)
}
\arguments{
\item{grQuery}{GRange object of query interval}

\item{bedFile}{BED file that is bgzipped and tabix'd}

\item{maxP}{maximum p-value of TFBS returned}
}
\value{
GRange of TFBS locations in p-values are scores
}
\description{
Read BED file of TFBS scores into GRange object
}
